Use loops and globs instead of :runtime
In older versions of Vim (prior to 7.3.1295), if a trigger contains
certain characters (like $), Vim can take a while to evaluate the globs.
Thus we avoid :runtime entirely in favor of s:Glob from before the
override branch was merged. This ends up being a bit slower in recent
Vim versions but much faster in older versions. I think this trade-off
is acceptable.

This also avoids tampering with 'runtimepath', which we never should
have been doing in the first place. That was probably a big reason the
override branch wasn't merged for a long time.

See also #195.