Fix bugs with completion and non-standard bindings
If the completion menu is visible and the user triggers a snippet
command, the completion menu is normally dismissed so that the snippet
command can execute. In order to do this, SnipMate uses `feedkeys` to
dismiss the menu, then also uses `feedkeys` so that the snippet function
gets called again after the menu is dismissed.

However, in the current SnipMate code, this second part is accomplished
via a call to `feedkeys("\<tab>")`, which only works if we're using the
default SnipMate bindings. If a non-standard binding is used, then the
snippet fails to trigger.

(This is perhaps less of an issue when the user has more direct control
over the completion menu, but it renders SnipMate unusable when it's
combined with a plugin like YouCompleteMe where the completion menu
appears automatically as you type.)

This change also fixes a possible issue where `feedkeys("\<tab>")` might
have triggered `snipMateNextOrTrigger` even if the original command the
user had executed might have been `snipMateTrigger`.